Add shift IntlTestingProvider
